# docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce Order Management System API
  description: Versioned REST API for products, orders, inventory and auth.
  version: "1.0.0"
servers:
  - url: "/api/v1"
    description: API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProductImport:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        uploader_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, processing, completed, failed]
        results_path:
          type: string
          nullable: true
        errors:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 1
        path: "imports/products/import.csv"
        uploader_id: 2
        status: "completed"
        results_path: "imports/products/import-results.json"
        errors: []
        created_at: "2025-11-18T12:00:00Z"
        updated_at: "2025-11-18T12:05:00Z"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
      example:
        message: "Validation failed"
        errors: { "email": ["The email field is required."] }
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
      example:
        id: 1
        name: "Admin User"
        email: "admin@example.com"
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        attributes:
          type: object
        price:
          type: number
          format: float
        quantity:
          type: integer
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        cost:
          type: number
          format: float
        quantity:
          type: integer
        low_stock_threshold:
          type: integer
        is_active:
          type: boolean
        vendor:
          $ref: '#/components/schemas/User'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
      example:
        id: 10
        name: "T-Shirt"
        sku: "TSHIRT-001"
        description: "100% cotton t-shirt"
        price: 19.99
        cost: 8.50
        quantity: 120
        low_stock_threshold: 10
        is_active: true
        vendor: { id: 2, name: "Vendor 1", email: "vendor@example.com" }
        variants: []
...
paths:
...
  /imports/{id}:
    get:
      tags: [Products]
      summary: Get product import status and results path
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product import status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductImport'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
...
  /products:
    get:
      tags: [Products]
      summary: List products (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: per_page
          schema: { type: integer }
        - in: query
          name: vendor_id
          schema: { type: integer }
      responses:
        '200':
          description: Paginated products
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
    post:
      tags: [Products]
      summary: Create a new product (vendor/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                sku: { type: string }
                description: { type: string }
                price: { type: number }
                cost: { type: number }
                quantity: { type: integer }
                low_stock_threshold: { type: integer }
                is_active: { type: boolean }
              required: [name, sku, price, quantity]
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          description: Forbidden
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { description: Not found }
    put:
      tags: [Products]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
    delete:
      tags: [Products]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '403': { description: Forbidden }
...
security:
  - bearerAuth: []